migrate volume tests and protocols to updated data types, fix assertions and replace int array checks with memory checks

+++ ControlPanel.Bridge/Device/DeviceProtocol/SetVolumeDeviceMessage.cs
@@ -6,5 +6,5 @@ namespace ControlPanel.Bridge.Device.DeviceProtocol;
 [MessagePackObject(true)]
 public record SetVolumeDeviceMessage(
     [property: Key("id")] AudioStreamId Id, 
-    [property: Key("volume")] double Volume)
+    [property: Key("volume")] float Volume)
     : DeviceMessage(MessageType.SetVolume), INotification;
